Remove util folder and move util.ts one level up
