+++ scenarioo-client/test/e2e/pages/searchResultsPage.ts
@@ -1,7 +1,7 @@
 'use strict';
 
 import {by, element, $, $$} from 'protractor';
-import * as Utils from '../util/util';
+import * as Utils from '../util';
 
 class SearchResultsPage {
