+++ scenarioo-client/test/e2e/pages/stepPage.ts
@@ -1,7 +1,7 @@
 'use strict';
 
 import {$, browser, by, element} from 'protractor';
-import * as Utils from '../util/util';
+import * as Utils from '../util';
 
 class StepPage {
