+++ scenarioo-client/test/e2e/pages/usecasePage.ts
@@ -1,7 +1,7 @@
 'use strict';
 
 import { by, ElementFinder, $ } from 'protractor';
-import * as Utils from '../util/util';
+import * as Utils from '../util';
 
 class UsecasePage {
