+++ scenarioo-client/test/e2e/specs/step_screen_annotations.ts
@@ -1,7 +1,7 @@
 'use strict';
 
 import { scenario, step, useCase } from 'scenarioo-js';
-import * as Utils from '../util/util';
+import * as Utils from '../util';
 import StepPage from '../pages/stepPage';
 
 useCase('Step - Screen Annotations')
