+++ scenarioo-client/test/e2e/specs/use_breadcrumbs.ts
@@ -1,7 +1,7 @@
 'use strict';
 
 import { scenario, step, useCase } from 'scenarioo-js';
-import * as Utils from '../util/util';
+import * as Utils from '../util';
 import HomePage from '../pages/homePage';
 import UsecasePage from '../pages/usecasePage';
 import BreadcrumbsPage from '../pages/breadcrumbsPage';
